kafka: consolidate offset commit handler
